Move to VSCode

+++ src/commands.h
@@ -5,7 +5,9 @@
 #ifndef SHAKE_COMMANDS_H
 #define SHAKE_COMMANDS_H
 
-int loadCommands(char ***cmds_out, char ***descs_out);
+#include <unistd.h>
+
+ssize_t loadCommands(char ***cmds_out, char ***descs_out);
 void loadCommandsSorted(void);
 
-#endif // SHAKE_COMMANDS_H
+#endif  // SHAKE_COMMANDS_H
