+++ src/runscripts/runscripts.c
@@ -2,51 +2,46 @@
 // Created by shayne on 11/9/16.
 //
 
+#include <glob.h>
+#include <stdlib.h>  // needed
 #include "../config.h"
 #include "../dbg.h"
 #include "../shakefile/shakefile.h"
-#include <glob.h>
-#include <stdlib.h> // needed
 
 #ifdef __APPLE__
 #include <libgen.h>
 #endif
 
-char *makescriptname(char *path)
-{
+char *makescriptname(char *path) {
     char *out = NULL;
     char *path_cp = NULL;
     char *path_base = basename(path_cp = strdup(path));
 
     int prefpos = config.cmd_prefix_len;
-    size_t cplen = strlen(path_base) - prefpos;
+    size_t cplen = strlen(path_base) - (size_t)prefpos;
     memcpy(&path_base[0], &path_base[prefpos], cplen);
     path_base[cplen] = '\0';
 
     char *dot = strrchr(path_base, '.');
     if (dot && dot != path_base) {
-        path_base[dot - path_base] = '\0'; // trim extension
+        path_base[dot - path_base] = '\0';  // trim extension
     }
 
     out = strdup(path_base);
     check_mem(out);
 
-error: // fallthrough
+error:  // fallthrough
     free(path_cp);
     return out;
 }
 
-int Runscripts_find_script(char *cmd_name, char **out)
-{
+int Runscripts_find_script(char *cmd_name, char **out) {
     int i;
     int rc;
     glob_t globbuf;
 
     char *pat = NULL;
-    rc = asprintf(&pat,
-                  "%s/%s/%s*",
-                  config.proj_dir,
-                  config.cmd_dir,
+    rc = asprintf(&pat, "%s/%s/%s*", config.proj_dir, config.cmd_dir,
                   config.cmd_prefix);
     check(rc > 0, "failed to format string");
 
@@ -55,8 +50,8 @@ int Runscripts_find_script(char *cmd_name, char **out)
 
     size_t cmd_name_l = strlen(cmd_name);
 
-    rc = 1; // default to failure
-    for (i = 0; i < globbuf.gl_pathc; i++) {
+    rc = 1;  // default to failure
+    for (i = 0; i < (int)globbuf.gl_pathc; i++) {
         char *gl_path = globbuf.gl_pathv[i];
         char *script_name = makescriptname(gl_path);
         check_debug(script_name != NULL, "failed to make script name");
@@ -74,7 +69,7 @@ int Runscripts_find_script(char *cmd_name, char **out)
         free(script_name);
     }
 
-error: // fallthrough
+error:  // fallthrough
     globfree(&globbuf);
     free(pat);
